+++ docs/changelog.rst
@@ -47,6 +47,7 @@ New features
 
 * Added new attribute :py:attr:`botogram.Message.pinned_message`
 * Added new attribute :py:attr:`botogram.Sticker.emoji`
+* Added new attribute :py:attr:`botogram.Chat.admins`
 * Added new method :py:meth:`botogram.Chat.leave`
 * Every method which sends something to a chat now returns the sent
   :py:class:`~botogram.Message`